Major redesign : New types of ChangeNodes added. Keyframes now belong to the timeline project. Models of the motion project have been dramatically changed(almost rewritten). Certain interfaces have been modified. Change handlers now have only one methods. The JavaFx project("editor") has mostly not been touched and is seemingly working fine.

+++ timeshift/util/src/main/java/com/nikhil/space/bezier/curve/CubicBezier.java
@@ -27,7 +27,7 @@ public class CubicBezier implements Parametric{
 	}
 
 	@Override
-	public UtilPoint getPointFor(float t) {
+	public UtilPoint getPointFor(double t) {
 		
 		//B(t)=(1-t)^3*P0 + 3*(1-t)^2*t*P1 + 3*(1-t)*t^2*P2 + t^3*P3
 		UtilPoint firstComponenet=starting.multiply((1-t)*(1-t)*(1-t));
